requirements.txt
langchain
langchain-community
langchain-huggingface
pypdf
python-dotenv
langchain-google-genai
